move towards cleaner models definitions
